+++ doc/reference/molecules/FileSelector.md
@@ -1,6 +1,6 @@
 <!-- 
 This is an auto-generated markdown. 
-You can change it in "src/organisms/FileSelector.jsx" and run build:docs to update this file.
+You can change it in "src/molecules/FileSelector.jsx" and run build:docs to update this file.
 -->
 # FileSelector
 The FileSelector component acts similar to <input type="file" /> but also provides useful tools to manage and display files.
@@ -8,7 +8,7 @@ The child should be a function which gets in passed following parameters:
 
 |Parameter | Type | Description |
 |----------|:----:|------------:|
-|`files`|Array<[File](https://developer.mozilla.org/en-US/docs/Web/API/File)>|Array with all added files (originating from [`FileList`](https://developer.mozilla.org/en-US/docs/Web/API/FileList))
+|`files`|Array<[File](https://developer.mozilla.org/en-US/docs/Web/API/File)>|Array containing added files (originating from [`FileList`](https://developer.mozilla.org/en-US/docs/Web/API/FileList))
 |`openDialog`|Function|Triggers the browsers file dialog
 |`getPreview`|Function|Get a preview generated by the browser (only works for images)
 | `removeFile`|Function|Removes a file from the list
@@ -35,8 +35,9 @@ The child should be a function which gets in passed following parameters:
 ## Usage
 | Name        | Type           | Description  |
 | ----------- |:--------------:| ------------:|
-|children|func|<br>Default: () => {}
+|children **(required)**|func|
 |name|string|Name of the field which is picked up by <Form> *
 |multiple|bool|Whether multiple files should be allowed *
 |required|bool|Whether this field is required *
 |accept|string|Accepted file types. See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#attr-accept *
+|onChange|func|Called when a file is added *
